% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.stat.R
\name{center.matrices}
\alias{center.matrices}
\title{Center a sample of matrices in Euclidean space}
\usage{
center.matrices(S, mu = NULL)
}
\arguments{
\item{S}{an array of matrices where the last dimension corresponds to sample size}

\item{mu}{the center}
}
\description{
Center a sample of matrices in Euclidean space
}
\keyword{internal}
