% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{ch.to.spd}
\alias{ch.to.spd}
\title{Map a Cholesky factor into its corresponding SPD}
\usage{
ch.to.spd(L)
}
\arguments{
\item{L}{a lower triangular matrix}
}
\value{
\code{L} multiplied by the transpose of \code{L}
}
\description{
Map a Cholesky factor into its corresponding SPD
}
\keyword{internal}
