% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{count.sym.matrix}
\alias{count.sym.matrix}
\title{count the number of matrices represented by an array or a list}
\usage{
count.sym.matrix(S)
}
\description{
count the number of matrices represented by an array or a list
}
\keyword{internal}
