% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{d.ch}
\alias{d.ch}
\title{the differential of ch at S}
\usage{
d.ch(S, W, L = NULL)
}
\arguments{
\item{S}{spd}

\item{W}{symmetric}

\item{L}{L=ch(S) if it is provided for speedup}
}
\description{
the differential of ch at S
}
\keyword{internal}
