% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.euclidean.R
\name{d.expm}
\alias{d.expm}
\title{Differential of matrix exponential of SPD manifold}
\usage{
d.expm(S, V)
}
\arguments{
\item{S}{base point}

\item{V}{tangent vector}
}
\description{
Differential of matrix exponential of SPD manifold
}
\keyword{internal}
