% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.euclidean.R
\name{d.logm}
\alias{d.logm}
\title{Differential of matrix log of SPD manifold}
\usage{
d.logm(S, V)
}
\arguments{
\item{S}{base point}

\item{V}{tangent vector}
}
\description{
Differential of matrix log of SPD manifold
}
\keyword{internal}
