% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{diag.part}
\alias{diag.part}
\title{The diagonal matrix of the diagonal part of a matrix}
\usage{
diag.part(A)
}
\description{
The diagonal matrix of the diagonal part of a matrix
}
\keyword{internal}
