% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{euclidean.mean}
\alias{euclidean.mean}
\title{Euclidean mean of a sample of matrices}
\usage{
euclidean.mean(S)
}
\arguments{
\item{S}{an array of matrices where the last dimension corresponds to sample size}
}
\description{
Euclidean mean of a sample of matrices
}
\keyword{internal}
