% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.stat.R
\name{frechet.mean.chol}
\alias{frechet.mean.chol}
\title{Frechet mean of positive lower-triangular matrix under Log-Cholesky metric}
\usage{
\method{frechet.mean}{chol}(S)
}
\description{
Frechet mean of positive lower-triangular matrix under Log-Cholesky metric
}
\keyword{internal}
