% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.interface.R
\name{frechet.mean}
\alias{frechet.mean}
\title{Frechet mean of matrices}
\usage{
frechet.mean(mfd, S)
}
\arguments{
\item{mfd}{an object created by \code{create.matrix.manifold}}

\item{S}{an \code{M*N*n} array of matrices or a list of \code{n} matrices, where \code{n} is the number of matrices}
}
\value{
the Frechet mean of the matrices in \code{S}
}
\description{
Frechet mean of matrices
}
