% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.euclidean.R
\name{frechet.mean.spd_LogEuclidean}
\alias{frechet.mean.spd_LogEuclidean}
\title{Frechet mean of matrices}
\usage{
\method{frechet.mean}{spd_LogEuclidean}(mfd, S)
}
\arguments{
\item{mfd}{an object created by \code{matrix.manifold}}

\item{S}{an \code{M*N*n} array of matrices or a list of \code{n} matrices, where \code{n} is the number of matrices}
}
\value{
the Frechet mean of the matrices in \code{S}
}
\description{
Frechet mean of matrices
}
\keyword{internal}
