% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{frobenius.norm}
\alias{frobenius.norm}
\title{Frobenius norm of a matrix}
\usage{
frobenius.norm(A)
}
\arguments{
\item{A}{a matrix}
}
\value{
the Frobenius norm of \code{A}
}
\description{
Frobenius norm of a matrix
}
\keyword{internal}
