% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{frobenius}
\alias{frobenius}
\title{Frobenius inner product of two matrices}
\usage{
frobenius(A, B = NULL)
}
\arguments{
\item{A}{a matrix}

\item{B}{a matrix with the same dimensions of \code{A}. If \code{B} is not provided, it is set to \code{A}}
}
\value{
the Frobenius inner product of \code{A} and \code{B}
}
\description{
Frobenius inner product of two matrices
}
\keyword{internal}
