% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.stat.R
\name{gen.lower}
\alias{gen.lower}
\title{generate a set of random lower triangular matrix of dimension d}
\usage{
gen.lower(d, n = 1, strict = F, mu = diag(d), sig = 1,
  dist = "Gaussian", drop = T)
}
\description{
generate a set of random lower triangular matrix of dimension d
}
\keyword{internal}
