% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{gen.sym}
\alias{gen.sym}
\title{generate a set of random symmetric matrix of dimension d}
\usage{
gen.sym(d, n = 1, mu = matrix(0, d, d), sig = 1, dist = "Gaussian",
  drop = T)
}
\arguments{
\item{mu}{the mean}

\item{sig}{a scalar representing the dispersion around the mean}

\item{dist}{the distribution, either 'uniform' or 'Gaussian'.}

\item{drop}{drop the last dimension if \code{n=1}}
}
\description{
generate a set of random symmetric matrix of dimension d
}
\keyword{internal}
