% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.interface.R
\name{gen.tangent.vectors}
\alias{gen.tangent.vectors}
\title{Generate a set of random matrices that represent tangent vectors at some point.}
\usage{
gen.tangent.vectors(mfd, n = 1, sig = 1, drop = T)
}
\arguments{
\item{mfd}{an object created by \code{create.matrix.manifold}}

\item{n}{sample size}

\item{sig}{the standard deviation of the normal distribution}
}
\value{
an \code{M*N*n} array of \code{n} matrices, where \code{M*N} is the dimensions of matrices
}
\description{
Generate a set of random matrices that represent tangent vectors at some point.
}
\details{
The generated samples have expectation zero and follow a isotropic D-dimensional normal distribution with isotropic variance \code{sig}, where D is the intrinsic dimension of the matrix manifold
}
