% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{gen.tangent.vectors.sym_Frobenius}
\alias{gen.tangent.vectors.sym_Frobenius}
\title{Generate a set of normally distributed random matrices that represent tangent vectors at some point.}
\usage{
\method{gen.tangent.vectors}{sym_Frobenius}(mfd, n = 1, sig = 1,
  drop = T)
}
\arguments{
\item{mfd}{an object created by \code{create.matrix.manifold}}

\item{n}{sample size}

\item{sig}{the standard deviation of the normal distribution}
}
\value{
an \code{M*N*n} array of \code{n} matrices, where \code{M*N} is the dimensions of matrices
}
\description{
Generate a set of normally distributed random matrices that represent tangent vectors at some point.
}
\keyword{internal}
