% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.stat.R
\name{gen.upper}
\alias{gen.upper}
\title{generate a random upper triangular matrix of dimension d}
\usage{
gen.upper(d, n = 1, strict = F, mu = diag(d), sig = 1,
  dist = "Gaussian", drop = T)
}
\description{
generate a random upper triangular matrix of dimension d
}
\keyword{internal}
