% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{geo.dist.chol}
\alias{geo.dist.chol}
\title{log-cholesky distance on the Cholesky factor space}
\usage{
\method{geo.dist}{chol}(L1, L2)
}
\description{
log-cholesky distance on the Cholesky factor space
}
\keyword{internal}
