% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.matrix.R
\name{geo.dist.ltpd_LogCholesky}
\alias{geo.dist.ltpd_LogCholesky}
\title{log-cholesky distance on the Cholesky factor space}
\usage{
\method{geo.dist}{ltpd_LogCholesky}(mfd, p, q, ...)
}
\arguments{
\item{p}{a LTPD}

\item{q}{a LTPD}
}
\description{
log-cholesky distance on the Cholesky factor space
}
\keyword{internal}
