% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.interface.R
\name{geo.dist}
\alias{geo.dist}
\title{Compute the geodesic distance}
\usage{
geo.dist(mfd, p, q, ...)
}
\arguments{
\item{mfd}{the manifold object created by \code{create.matrix.manifold}}

\item{p}{a point on the manifold}

\item{q}{a point on the manifold}

\item{...}{other parameters (primarily for computation speedup)}
}
\value{
the geodesic distance between \code{p} and \code{q}
}
\description{
Compute the geodesic distance
}
