% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{geo.dist.spd_LogCholesky}
\alias{geo.dist.spd_LogCholesky}
\title{Compute the geodesic distance of two SPD in Log-Cholesky metric}
\usage{
\method{geo.dist}{spd_LogCholesky}(mfd, S1, S2, opt.param = NULL)
}
\arguments{
\item{S1}{a SPD}

\item{S2}{a SPD}

\item{opt.param}{optional parameters, primarily for speedup. In the case log-Cholesky, it can be ch(S1) and ch(S2)}
}
\description{
Compute the geodesic distance of two SPD in Log-Cholesky metric
}
\keyword{internal}
