% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.euclidean.R
\name{geo.dist.spd_LogEuclidean}
\alias{geo.dist.spd_LogEuclidean}
\title{Compute the geodesic distance}
\usage{
\method{geo.dist}{spd_LogEuclidean}(mfd, p, q, ...)
}
\arguments{
\item{mfd}{the manifold object created by \code{matrix.manifold}}

\item{p}{a point on the manifold}

\item{q}{a point on the manifold}

\item{...}{other parameters (primarily for computation speedup)}
}
\value{
the geodesic distance between \code{p} and \code{q}
}
\description{
Compute the geodesic distance
}
\keyword{internal}
