% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{geo.dist.sym_Frobenius}
\alias{geo.dist.sym_Frobenius}
\title{Compute the geodesic distance of two symmetric in Forbenius metric}
\usage{
\method{geo.dist}{sym_Frobenius}(mfd, S1, S2, opt.param = NULL)
}
\arguments{
\item{S1}{a symmetric}

\item{S2}{a symmetric}

\item{opt.param}{optional parameters, primarily for speedup.}
}
\description{
Compute the geodesic distance of two symmetric in Forbenius metric
}
\keyword{internal}
