% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.matrix.R
\name{geodesic.ltpd_LogCholesky}
\alias{geodesic.ltpd_LogCholesky}
\title{geodesic on the Cholesky factor space}
\usage{
\method{geodesic}{ltpd_LogCholesky}(mfd, p, u, t, ...)
}
\arguments{
\item{p}{starting point}

\item{u}{direction}

\item{t}{time}
}
\description{
geodesic on the Cholesky factor space
}
\keyword{internal}
