% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.interface.R
\name{geodesic}
\alias{geodesic}
\title{Compute the geodesic starting at some point on a manifold}
\usage{
geodesic(mfd, p, u, t, ...)
}
\arguments{
\item{mfd}{the manifold object created by \code{create.matrix.manifold}}

\item{p}{the starting point of the geodesic}

\item{u}{a tangent vector at \code{p}, representing the direction of the geodesic}

\item{t}{a vector of nonnegative real numbers}

\item{...}{other parameters (primarily for computation speedup)}
}
\value{
an array of matrices which are the geodesic evaluated at \code{t}. The last dimension of the array corresponds to \code{t}
}
\description{
Compute the geodesic starting at some point on a manifold
}
