% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{geodesic.spd_LogCholesky}
\alias{geodesic.spd_LogCholesky}
\title{Log-Cholesky geodesic of SPD at S}
\usage{
\method{geodesic}{spd_LogCholesky}(mfd, p, u, t, ...)
}
\arguments{
\item{p}{spd}

\item{u}{symmetric}

\item{t}{time}

\item{...}{L=ch(S)}
}
\description{
Log-Cholesky geodesic of SPD at S
}
\keyword{internal}
