% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{geodesic.sym_Frobenius}
\alias{geodesic.sym_Frobenius}
\title{Frobenius geodesic of symmetric matrices}
\usage{
\method{geodesic}{sym_Frobenius}(mfd, p, u, t, ...)
}
\arguments{
\item{mfd}{the manifold object}

\item{p}{symmetric}

\item{u}{symmetric}

\item{t}{time}

\item{...}{optional parameters, primarily for speedup. Not used here.}
}
\description{
Frobenius geodesic of symmetric matrices
}
\keyword{internal}
