% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{half}
\alias{half}
\title{the half operation of A
return a matrix B of the same dimension of A, such that B(i,j)=A(i,j) if i>j, B(i,i)=A(i,i)/2, and B(i,j)=0 if i<j}
\usage{
half(A)
}
\description{
the half operation of A
return a matrix B of the same dimension of A, such that B(i,j)=A(i,j) if i>j, B(i,i)=A(i,i)/2, and B(i,j)=0 if i<j
}
\keyword{internal}
