% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{is.spd}
\alias{is.spd}
\title{Check whether a matrix is a symmetric positive definite matrix}
\usage{
is.spd(A)
}
\arguments{
\item{A}{a matrix}
}
\value{
TRUE or FALSE depending on whether \code{A} is an SPD
}
\description{
Check whether a matrix is a symmetric positive definite matrix
}
\keyword{internal}
