% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{is.sym.in.matrix.form}
\alias{is.sym.in.matrix.form}
\title{Check whether a matrix is represented by its vector form}
\usage{
is.sym.in.matrix.form(S)
}
\description{
Check whether a matrix is represented by its vector form
}
\keyword{internal}
