% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{is.sym}
\alias{is.sym}
\title{Check whether a matrix is a symmetric matrix}
\usage{
is.sym(A)
}
\arguments{
\item{A}{a matrix}
}
\value{
TRUE or FALSE depending on whether \code{A} is symmetric
}
\description{
Check whether a matrix is a symmetric matrix
}
\keyword{internal}
