% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{lower.part}
\alias{lower.part}
\title{return the lower triangular part of A}
\usage{
lower.part(A, strict = T)
}
\arguments{
\item{strict}{whether a strict triangular matrix is returned, i.e. whether the diagonal elements are EXCLUDED (strict) or not (non-strict)}
}
\description{
return the lower triangular part of A
}
\keyword{internal}
