% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.matrix.R
\name{lower.to.vec.atomic}
\alias{lower.to.vec.atomic}
\title{transform a (set of) lower triangular matrix into a vector representation, diagonal by diagonal, starting from the main diagonal}
\usage{
lower.to.vec.atomic(L)
}
\description{
transform a (set of) lower triangular matrix into a vector representation, diagonal by diagonal, starting from the main diagonal
}
\keyword{internal}
