% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{lower.to.vec}
\alias{lower.to.vec}
\title{transform a (set of) lower triangular matrix into a vector representation, diagonal by diagonal, starting from the main diagonal}
\usage{
lower.to.vec(L, drop = F)
}
\description{
transform a (set of) lower triangular matrix into a vector representation, diagonal by diagonal, starting from the main diagonal
}
\keyword{internal}
