% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{make.sym}
\alias{make.sym}
\title{Make a matrix into a symmetric matrix to correct potential asymmetry due to finite-precision operations}
\usage{
make.sym(S)
}
\description{
Make a matrix into a symmetric matrix to correct potential asymmetry due to finite-precision operations
}
\keyword{internal}
