% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.interface.R
\name{matrix.manifold}
\alias{matrix.manifold}
\title{Create an object that represent a Riemannian structure on a matrix space}
\usage{
matrix.manifold(manifold, metric, dim)
}
\arguments{
\item{manifold}{matrix manifold, one of 'lt', 'spd','stiefel' and 'sym'}

\item{metric}{Riemannian metric, 'LogCholesky','AffineInvariant','LogEuclidean' for spd}

\item{dim}{the dimensions of the matrices, must be an positive integer or a vector of two positive integers}
}
\value{
an object containing essential information about the defined matrix manifold
}
\description{
Create an object that represent a Riemannian structure on a matrix space
}
