% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{metric.chol}
\alias{metric.chol}
\title{log-cholesky metric on the Cholesky factor space}
\usage{
metric.chol(L, X, Y)
}
\description{
log-cholesky metric on the Cholesky factor space
}
\keyword{internal}
