% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.matrix.R
\name{parallel.transport.ltpd_LogCholesky}
\alias{parallel.transport.ltpd_LogCholesky}
\title{Parallel transport of a tangent vector from a point to another}
\usage{
\method{parallel.transport}{ltpd_LogCholesky}(mfd, p, q, v, ...)
}
\arguments{
\item{p}{a LTPD matrix}

\item{q}{a LTPD matrix}

\item{v}{a tangent vector at \code{p}}
}
\description{
Parallel transport of a tangent vector from a point to another
}
\keyword{internal}
