% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{parallel.transport.spd_LogCholesky}
\alias{parallel.transport.spd_LogCholesky}
\title{Parallel transport of a tangent vector from a point to another}
\usage{
\method{parallel.transport}{spd_LogCholesky}(mfd, S1, S2, W)
}
\arguments{
\item{S1}{a SPD}

\item{S2}{a SPD}

\item{W}{a tangent vector at \code{S1}}
}
\description{
Parallel transport of a tangent vector from a point to another
}
\keyword{internal}
