% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.euclidean.R
\name{parallel.transport.spd_LogEuclidean}
\alias{parallel.transport.spd_LogEuclidean}
\title{Parallel transport of a tangent vector from one point to another along the geodesic between the two points}
\usage{
\method{parallel.transport}{spd_LogEuclidean}(mfd, p, q, v, ...)
}
\arguments{
\item{mfd}{the manifold object created by \code{matrix.manifold}}

\item{p}{a point on the manifold}

\item{q}{a point on the manifold}

\item{v}{a tangent vector at \code{p}}

\item{...}{other parameters (primarily for computation speedup)}
}
\value{
a tangent vector at \code{q}
}
\description{
Parallel transport of a tangent vector from one point to another along the geodesic between the two points
}
\keyword{internal}
