% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{parallel.transport.sym_Frobenius}
\alias{parallel.transport.sym_Frobenius}
\title{Parallel transport of a tangent vector from a point to another}
\usage{
\method{parallel.transport}{sym_Frobenius}(mfd, S1, S2, W)
}
\arguments{
\item{S1}{symmetric}

\item{S2}{symmetric}

\item{W}{a tangent vector at \code{S1}}
}
\description{
Parallel transport of a tangent vector from a point to another
}
\keyword{internal}
