% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{rie.exp.chol}
\alias{rie.exp.chol}
\title{log-cholesky exponential map on the Cholesky factor space}
\usage{
\method{rie.exp}{chol}(L, X)
}
\description{
log-cholesky exponential map on the Cholesky factor space
}
\keyword{internal}
