% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.matrix.R
\name{rie.exp.ltpd_LogCholesky}
\alias{rie.exp.ltpd_LogCholesky}
\title{log-cholesky exponential map on the Cholesky factor space}
\usage{
\method{rie.exp}{ltpd_LogCholesky}(mfd, p, v, ...)
}
\arguments{
\item{p}{the base point}

\item{v}{a tangent vector at p}
}
\description{
log-cholesky exponential map on the Cholesky factor space
}
\keyword{internal}
