% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{rie.exp.spd_LogCholesky}
\alias{rie.exp.spd_LogCholesky}
\title{Riemannian exponential map}
\usage{
\method{rie.exp}{spd_LogCholesky}(mfd, p, v, ...)
}
\arguments{
\item{p}{the foot point of the exp map}

\item{v}{a tangent vector at \code{S}}

\item{...}{L=ch(S)}
}
\description{
Riemannian exponential map
}
\keyword{internal}
