% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.euclidean.R
\name{rie.exp.spd_LogEuclidean}
\alias{rie.exp.spd_LogEuclidean}
\title{Compute the Riemannian exponential map}
\usage{
\method{rie.exp}{spd_LogEuclidean}(mfd, p, v, ...)
}
\arguments{
\item{mfd}{the manifold object created by \code{matrix.manifold}}

\item{p}{the foot point of the exponential map}

\item{v}{a tangent vector at \code{p}}

\item{...}{other parameters (primarily for computation speedup)}
}
\value{
a matrix that is the exponential map of \code{v}
}
\description{
Compute the Riemannian exponential map
}
\keyword{internal}
