% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{rie.log.chol}
\alias{rie.log.chol}
\title{log-cholesky log map on the Cholesky factor space}
\usage{
\method{rie.log}{chol}(L1, L2)
}
\description{
log-cholesky log map on the Cholesky factor space
}
\keyword{internal}
