% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.matrix.R
\name{rie.log.ltpd_LogCholesky}
\alias{rie.log.ltpd_LogCholesky}
\title{log-cholesky log map on the Cholesky factor space}
\usage{
\method{rie.log}{ltpd_LogCholesky}(mfd, p, q, ...)
}
\arguments{
\item{p}{the foot point of the log map}

\item{q}{the point to be mapped}
}
\description{
log-cholesky log map on the Cholesky factor space
}
\keyword{internal}
