% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.interface.R
\name{rie.log}
\alias{rie.log}
\title{Compute the Riemannian logarithm map}
\usage{
rie.log(mfd, p, q, ...)
}
\arguments{
\item{mfd}{the manifold object created by \code{create.matrix.manifold}}

\item{p}{the foot point of the logarithm map}

\item{q}{a point on the manifold}

\item{...}{other parameters (primarily for computation speedup)}
}
\value{
a matrix that is the logarithm map of \code{q}
}
\description{
Compute the Riemannian logarithm map
}
