% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{rie.log.spd_LogCholesky}
\alias{rie.log.spd_LogCholesky}
\title{Riemannian logarithmic map}
\usage{
\method{rie.log}{spd_LogCholesky}(mfd, p, q, ...)
}
\arguments{
\item{p}{the foot point of the log map}

\item{q}{the point to be mapped}

\item{...}{L1=ch(p) and L2=ch(q)}
}
\description{
Riemannian logarithmic map
}
\keyword{internal}
