% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{rie.log.sym_Frobenius}
\alias{rie.log.sym_Frobenius}
\title{Frobenius logarithmic map of symmetric matrices at S}
\usage{
\method{rie.log}{sym_Frobenius}(mfd, S, Q, ...)
}
\arguments{
\item{S}{the foot point of the log map}

\item{Q}{a tangent vector at \code{S}}

\item{...}{other parameters (primarily for computation speedup)}
}
\value{
a matrix that is the log map of \code{Q}
}
\description{
Frobenius logarithmic map of symmetric matrices at S
}
