% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.matrix.R
\name{rie.metric.ltpd_LogCholesky}
\alias{rie.metric.ltpd_LogCholesky}
\title{log-cholesky metric on the Cholesky factor space}
\usage{
\method{rie.metric}{ltpd_LogCholesky}(mfd, p, u, v, ...)
}
\arguments{
\item{p}{the base point}

\item{u}{a tangent vector at p}

\item{v}{a tangent vector at p}
}
\description{
log-cholesky metric on the Cholesky factor space
}
\keyword{internal}
