% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{rie.metric.spd_LogCholesky}
\alias{rie.metric.spd_LogCholesky}
\title{Log-Cholesky metric of SPD at S}
\usage{
\method{rie.metric}{spd_LogCholesky}(mfd, S, W, V, opt.param = NULL)
}
\arguments{
\item{S}{spd, the base point}

\item{W}{symmetric, a tangent vector at S}

\item{V}{symmetric, a tangent vector at S}

\item{opt.param}{optional parameters, primarily for speedup. In the case log-Cholesky, it can be ch(S)}
}
\description{
Log-Cholesky metric of SPD at S
}
\keyword{internal}
