% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.euclidean.R
\name{rie.metric.spd_LogEuclidean}
\alias{rie.metric.spd_LogEuclidean}
\title{Compute the Riemannian metric}
\usage{
\method{rie.metric}{spd_LogEuclidean}(mfd, p, u, v, ...)
}
\arguments{
\item{mfd}{the manifold object created by \code{matrix.manifold}}

\item{p}{the foot point of the metric}

\item{u}{a tangent vector at \code{p}}

\item{v}{a tangent vector at \code{q}}

\item{...}{other parameters (primarily for computation speedup)}
}
\value{
a number that is the Riemannian metric of \code{u} and \code{v}
}
\description{
Compute the Riemannian metric
}
\keyword{internal}
