% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{rie.metric.sym_Frobenius}
\alias{rie.metric.sym_Frobenius}
\title{Frobenius metric of symmetric matrices at S}
\usage{
\method{rie.metric}{sym_Frobenius}(mfd, S, W, V, opt.param = NULL)
}
\arguments{
\item{S}{symmetric, the base point}

\item{W}{symmetric, a tangent vector at S}

\item{V}{symmetric, a tangent vector at S}

\item{opt.param}{optional parameters, primarily for speedup. Not used here}
}
\description{
Frobenius metric of symmetric matrices at S
}
\keyword{internal}
