% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.matrix.R
\name{rltmatrix}
\alias{rltmatrix}
\title{generate a set of random lower triangular matrix of dimension d}
\usage{
rltmatrix(
  d,
  n = 1,
  strict = F,
  mu = NULL,
  sig = 1,
  dist = "Gaussian",
  drop = T
)
}
\description{
generate a set of random lower triangular matrix of dimension d
}
\keyword{internal}
