% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.matrix.R
\name{rmatrix.ltpd_LogCholesky}
\alias{rmatrix.ltpd_LogCholesky}
\title{Generate a set of random matrices on a matrix manifold}
\usage{
\method{rmatrix}{ltpd_LogCholesky}(mfd, n = 1, mu = NULL, sig = 1, drop = T)
}
\arguments{
\item{mfd}{an object created by \code{matrix.manifold}}

\item{n}{sample size}

\item{mu}{the Frechet mean. If \code{NULL} is given, then it is the identity element, e.g., for SPD, it is the identity matrix}

\item{sig}{the standard deviation of the normal distribution}

\item{drop}{whether return the result as a matrix when \code{n=1}}
}
\value{
an \code{M*N*n} array of \code{n} matrices, where \code{M*N} is the dimensions of matrices
}
\description{
Generate a set of random matrices on a matrix manifold
}
\details{
The generated samples have Frechet mean \code{mu}. The logarithmic maps of these samples at \code{mu} follow a isotropic D-dimensional normal distribution with isotropic variance \code{sig}, where D is the intrinsic dimension of the matrix manifold
}
\keyword{internal}
