% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.euclidean.R
\name{rtvecor.spd_LogEuclidean}
\alias{rtvecor.spd_LogEuclidean}
\title{Generate a set of random matrices that represent tangent vectors at some point.}
\usage{
\method{rtvecor}{spd_LogEuclidean}(mfd, n = 1, sig = 1, drop = T)
}
\arguments{
\item{mfd}{an object created by \code{matrix.manifold}}

\item{n}{sample size}

\item{sig}{the standard deviation of the normal distribution}

\item{drop}{whether return the result as a matrix when \code{n=1}}
}
\value{
an \code{M*N*n} array of \code{n} matrices, where \code{M*N} is the dimensions of matrices
}
\description{
Generate a set of random matrices that represent tangent vectors at some point.
}
\details{
The generated samples have expectation zero and follow a isotropic D-dimensional normal distribution with isotropic variance \code{sig}, where D is the intrinsic dimension of the matrix manifold
}
\keyword{internal}
