% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangular.matrix.R
\name{rutmatrix}
\alias{rutmatrix}
\title{generate a random upper triangular matrix of dimension d}
\usage{
rutmatrix(
  d,
  n = 1,
  strict = F,
  mu = diag(d),
  sig = 1,
  dist = "Gaussian",
  drop = T
)
}
\description{
generate a random upper triangular matrix of dimension d
}
\keyword{internal}
