% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{strict.lower.part}
\alias{strict.lower.part}
\title{The strict lower triangular part of a matrix}
\usage{
strict.lower.part(A)
}
\description{
The strict lower triangular part of a matrix
}
\keyword{internal}
