% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{strict.upper.part}
\alias{strict.upper.part}
\title{The strict upper triangular part of a matrix}
\usage{
strict.upper.part(A)
}
\description{
The strict upper triangular part of a matrix
}
\keyword{internal}
