% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{sym.to.matrix.form}
\alias{sym.to.matrix.form}
\title{transform a matrix into its matrix form}
\usage{
sym.to.matrix.form(S)
}
\description{
transform a matrix into its matrix form
}
\keyword{internal}
