% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{sym.to.vec.atomic}
\alias{sym.to.vec.atomic}
\title{Transform a lower triangular matrix into a vector representation, diagonal by diagonal, starting from the main diagonal}
\usage{
sym.to.vec.atomic(S)
}
\description{
Transform a lower triangular matrix into a vector representation, diagonal by diagonal, starting from the main diagonal
}
\keyword{internal}
