% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{upper.part}
\alias{upper.part}
\title{The upper triangular part of a matrix}
\usage{
upper.part(A, strict = T)
}
\arguments{
\item{strict}{whether a strict triangular matrix is returned, i.e. whether the diagonal elements are EXCLUDED (strict) or not (non-strict)}
}
\description{
The upper triangular part of a matrix
}
\keyword{internal}
