% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{vec.to.lower.atomic}
\alias{vec.to.lower.atomic}
\title{recover a lower triangular matrix from its vector representation}
\usage{
vec.to.lower.atomic(v)
}
\description{
recover a lower triangular matrix from its vector representation
}
\keyword{internal}
