% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{vec.to.sym.atomic}
\alias{vec.to.sym.atomic}
\title{Transform a vector created by \code{sym.to.vec.atomic} or \code{sym.to.vec} into a symmetric matric}
\usage{
vec.to.sym.atomic(v)
}
\description{
Transform a vector created by \code{sym.to.vec.atomic} or \code{sym.to.vec} into a symmetric matric
}
\keyword{internal}
