% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{vec.to.sym}
\alias{vec.to.sym}
\title{recover a (set of) lower triangular matrix from its vector representation}
\usage{
vec.to.sym(v, drop = F)
}
\description{
recover a (set of) lower triangular matrix from its vector representation
}
\keyword{internal}
