% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{geo.dist.spd_LogCholesky}
\alias{geo.dist.spd_LogCholesky}
\title{Compute the geodesic distance of two SPD in Log-Cholesky metric}
\usage{
\method{geo.dist}{spd_LogCholesky}(mfd, p, q, ...)
}
\arguments{
\item{p}{a SPD}

\item{q}{a SPD}

\item{...}{L1=ch(S1) and L2=ch(S2)}
}
\description{
Compute the geodesic distance of two SPD in Log-Cholesky metric
}
\keyword{internal}
