% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{geo.dist.sym_Frobenius}
\alias{geo.dist.sym_Frobenius}
\title{Compute the geodesic distance of two symmetric in Forbenius metric}
\usage{
\method{geo.dist}{sym_Frobenius}(mfd, p, q, ...)
}
\arguments{
\item{mfd}{the manifold object}

\item{p}{a symmetric}

\item{q}{a symmetric}

\item{...}{optional parameters, primarily for speedup.}
}
\description{
Compute the geodesic distance of two symmetric in Forbenius metric
}
\keyword{internal}
