% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{geodesic.spd_LogCholesky}
\alias{geodesic.spd_LogCholesky}
\title{Log-Cholesky geodesic of SPD at S}
\usage{
\method{geodesic}{spd_LogCholesky}(mfd, S, W, t, opt.param = NULL)
}
\arguments{
\item{S}{spd}

\item{W}{symmetric}

\item{t}{time}

\item{opt.param}{optional parameters, primarily for speedup. In the case log-Cholesky, it can be ch(S)}
}
\description{
Log-Cholesky geodesic of SPD at S
}
\keyword{internal}
