% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{geodesic.sym_Frobenius}
\alias{geodesic.sym_Frobenius}
\title{Frobenius geodesic of symmetric matrices at S}
\usage{
\method{geodesic}{sym_Frobenius}(mfd, S, W, t, opt.param = NULL)
}
\arguments{
\item{S}{symmetric}

\item{W}{symmetric}

\item{t}{time}

\item{opt.param}{optional parameters, primarily for speedup. Not used here.}
}
\description{
Frobenius geodesic of symmetric matrices at S
}
\keyword{internal}
