% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{parallel.transport.spd_LogCholesky}
\alias{parallel.transport.spd_LogCholesky}
\title{Parallel transport of a tangent vector from a point to another}
\usage{
\method{parallel.transport}{spd_LogCholesky}(mfd, p, q, v, ...)
}
\arguments{
\item{p}{a SPD}

\item{q}{a SPD}

\item{v}{a tangent vector at \code{S1}}
}
\description{
Parallel transport of a tangent vector from a point to another
}
\keyword{internal}
