% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{rie.exp.sym_Frobenius}
\alias{rie.exp.sym_Frobenius}
\title{Frobenius exponential map of symmetric matrices at S}
\usage{
\method{rie.exp}{sym_Frobenius}(mfd, S, V, ...)
}
\arguments{
\item{S}{the foot point of the exponential map}

\item{V}{a tangent vector at \code{S}}

\item{...}{other parameters (primarily for computation speedup)}
}
\value{
a matrix that is the exponential map of \code{V}
}
\description{
Frobenius exponential map of symmetric matrices at S
}
