% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{rie.log.sym_Frobenius}
\alias{rie.log.sym_Frobenius}
\title{Frobenius logarithmic map of symmetric matrices at p}
\usage{
\method{rie.log}{sym_Frobenius}(mfd, p, q, ...)
}
\arguments{
\item{mfd}{the manifold object}

\item{p}{the foot point of the log map}

\item{q}{a tangent vector at \code{p}}

\item{...}{other parameters (primarily for computation speedup)}
}
\value{
a matrix that is the log map of \code{q}
}
\description{
Frobenius logarithmic map of symmetric matrices at p
}
\keyword{internal}
