% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.cholesky.R
\name{rie.metric.spd_LogCholesky}
\alias{rie.metric.spd_LogCholesky}
\title{Log-Cholesky metric of SPD at p}
\usage{
\method{rie.metric}{spd_LogCholesky}(mfd, p, u, v, ...)
}
\arguments{
\item{p}{spd, the base point}

\item{u}{symmetric, a tangent vector at p}

\item{v}{symmetric, a tangent vector at p}

\item{...}{L=ch(p)}
}
\description{
Log-Cholesky metric of SPD at p
}
\keyword{internal}
