% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symmetric.matrix.R
\name{rie.metric.sym_Frobenius}
\alias{rie.metric.sym_Frobenius}
\title{Frobenius metric of symmetric matrices at p}
\usage{
\method{rie.metric}{sym_Frobenius}(mfd, p, u, v, ...)
}
\arguments{
\item{p}{symmetric, the base point}

\item{u}{symmetric, a tangent vector at p}

\item{v}{symmetric, a tangent vector at p}

\item{...}{optional parameters, primarily for speedup. Not used here}
}
\description{
Frobenius metric of symmetric matrices at p
}
\keyword{internal}
