% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.utility.R
\name{vec.to.lower}
\alias{vec.to.lower}
\title{recover a (set of) lower triangular matrix from its vector representation}
\usage{
vec.to.lower(v, drop = F)
}
\description{
recover a (set of) lower triangular matrix from its vector representation
}
\keyword{internal}
